end the inital work on admin panel

+++ admin/categories.php
@@ -1,23 +1,24 @@
 <?php
 session_start();
+if(!isset($_SESSION['user']))
+    redirect();
 $title = 'Categories Manage';
 include 'init.php';
-if(!isset($_SESSION['user']))
-    redirect(1);
 $arr = array('ASC', 'DESC');
 $sort = (isset($_GET['sort']) && in_array($_GET['sort'], $arr)) ? test_input($_GET['sort']) : 'ASC';
 $rows = selectRecords('*', 'categories', null, 'id', $sort);
 ?>
 <h1 class="text-center">Categories Management</h1>
 <div class="container cats">
+    <?php if (isset($_GET['message'])) echo '<div class="alert alert-warning">' . $_GET['message'] . '</div>'; ?>
     <a href="addCategory.php" class="btn btn-primary"><i class="fa fa-plus"></i> New Category</a><br><br>
     <div class="card">
         <div class="card-header">
             Categories
             <div class="options pull-right">
                 <i class="fa fa-sort"></i> Ordering: [
-                <a <?php if($sort == 'ASC') { echo 'class="active"'; } ?> href="<?php htmlspecialchars($_SERVER['PHP_SELF']) ?>?sort=ASC"><i class="fa fa-sort-alpha-asc"></i></a> |
-                <a <?php if($sort == 'DESC') { echo 'class="active"'; } ?> href="<?php htmlspecialchars($_SERVER['PHP_SELF']) ?>?sort=DESC"><i class="fa fa-sort-alpha-desc"></i></a>]
+                <a <?php if($sort == 'ASC') echo 'class="active"'; ?> href="<?php htmlspecialchars($_SERVER['PHP_SELF']) ?>?sort=ASC"><i class="fa fa-sort-alpha-asc"></i></a> |
+                <a <?php if($sort == 'DESC') echo 'class="active"'; ?> href="<?php htmlspecialchars($_SERVER['PHP_SELF']) ?>?sort=DESC"><i class="fa fa-sort-alpha-desc"></i></a>]
                 View: [
                 <span class="active" data-view="full">Full</span> |
                 <span>Classic</span>]
@@ -29,8 +30,8 @@ $rows = selectRecords('*', 'categories', null, 'id', $sort);
             <?php } else { foreach($rows as $row) { ?>
                 <div class="cat">
                     <div class="hidden-btn">
-                        <a class="btn btn-info" href="editCategory.php?id=<?php echo $row['id'] ?>"><i class="fa fa-edit"></i>Edit</a>
-                        <a class="btn btn-danger" href="deleteCategory.php?id=<?php echo $row['id'] ?>"><i class="fa fa-close"></i>Delete</a>
+                        <a class="btn btn-sm btn-info" href="editCategory.php?id=<?php echo $row['id'] ?>"><i class="fa fa-edit"></i>Edit</a>
+                        <a class="btn btn-sm btn-danger confirm-delete" href="deleteCategory.php?id=<?php echo $row['id'] ?>"><i class="fa fa-close"></i>Delete</a>
                     </div>
                     <h3><?php echo $row['title'] ?></h3>
                     <div class="view">
